linux/requirements.txt
PyGObject
python-dotenv
websockets
pyinstaller
